perf(e2e): cut redundant specs and CI setup work
The browser suite spent most of its time re-driving the full
Slack -> implement -> PR flow to produce fixtures it only read, and each
of the three CI shards independently reinstalled Chromium and rebuilt the
dashboard.

- Drop two specs fully covered elsewhere: the full_flow "no mention"
  case (duplicates slack_untagged_reply, and proved it with a 3s sleep)
  and the sandbox_id shared-thread case (sidebar visibility is covered by
  ownerless, the copy action by the desktop case in the same file).
- Share one fixture thread across the read-only transcript assertions and
  merge the folded-work and inline-diff tests, which used identical setup.
- Split dashboard.spec.ts, 45% of the suite, into transcript / pull
  request / handoff files so file-level sharding can balance them.
- Shorten the scripted subagent sleep from 12s to 3s and the busy-hold
  fallback from 20s to 8s.
- Drop the global expect timeout from 60s to 20s so a real failure
  reports in seconds instead of minutes.
- Record traces and video only for failures locally; E2E_ARTIFACTS=1
  restores always-on capture.
- Cache Chromium and ui/.output across the CI shards.
