+++ tests/e2e/tests/dashboard.spec.ts
@@ -1,193 +1,21 @@
-import { test, expect, type Page } from "@playwright/test";
+import { test, expect } from "@playwright/test";
+import {
+  OTHER_USER,
+  SAME_USER,
+  composerFor,
+  expectTranscriptVisible,
+  loginAs,
+  openRunningThreadViaSlackLink,
+  openThreadViaSlackLink,
+  threadIdFromUrl,
+  threadState,
+  typeIntoComposer,
+  waitForStateToContain,
+  waitForThreadIdle,
+} from "./helpers/dashboard";
 
 // Drives the REAL built ui/ app (served same-origin from the harness) for the
 // Slack → web handoff. Only the LLM/GitHub/Slack/token boundaries are faked.
-const SAME_USER = { login: "alice", email: "alice@example.com" };
-const OTHER_USER = { login: "bob", email: "bob@example.com" };
-
-async function loginAs(page: Page, user: { login: string; email: string }) {
-  const res = await page.request.post("/control/login", { data: user });
-  expect(res.ok()).toBeTruthy();
-}
-
-// The composer is a rich-text editor, not a <textarea>: it carries the prompt
-// as `aria-placeholder` plus a visible overlay, so `getByPlaceholder` (which
-// only matches the `placeholder` attribute) can't see it. Assert on both hooks
-// so the visible prompt text stays covered.
-function composerFor(page: Page, placeholder: RegExp) {
-  return {
-    editor: page.getByTestId("composer-editor"),
-    prompt: page.getByText(placeholder),
-  };
-}
-
-// Typing goes through real key events rather than `fill()`: the editor builds
-// its state from beforeinput/keydown, and `fill()`'s single bulk insert leaves
-// it out of sync with the DOM.
-async function typeIntoComposer(page: Page, text: string) {
-  const editor = page.getByTestId("composer-editor");
-  await editor.click();
-  await editor.pressSequentially(text);
-  await editor.press("Enter");
-}
-
-async function setRepoPrivate(page: Page, value: boolean) {
-  const res = await page.request.post("/control/repo-private", {
-    data: { private: value },
-  });
-  expect(res.ok()).toBeTruthy();
-}
-
-async function setPullRequestHealth(
-  page: Page,
-  values: Record<string, unknown>,
-) {
-  const res = await page.request.post("/control/pull-request-health", {
-    data: { number: 1, ...values },
-  });
-  expect(res.ok()).toBeTruthy();
-}
-
-// Hold the fake run open long enough to load its busy composer and queue a
-// follow-up. It may finish while the UI observes the next server refresh.
-async function openRunningThreadViaSlackLink(page: Page) {
-  await page.goto("/mock/slack");
-  await page.locator("#reset").click();
-  await expect(page.locator("#thread")).toContainText("No messages yet");
-  await page
-    .locator("#text")
-    .fill("<@U0BOT> E2E_BUSY_HOLD:8 please add a greet() helper and open a PR");
-  await page.locator("#send").click();
-
-  const webLink = page.locator('.msg.bot a[href*="/agents/"]').first();
-  await expect(webLink).toBeVisible();
-  await webLink.click();
-  await expect(page).toHaveURL(/\/agents\//);
-}
-
-// Run the Slack flow so a thread + PR exist, then click the bot's real
-// "Open in Web" link, landing on the actual dashboard app.
-async function openThreadViaSlackLink(
-  page: Page,
-  options: { repoPrivate?: boolean; message?: string } = {},
-) {
-  await page.goto("/mock/slack");
-  await page.locator("#reset").click();
-  if (options.repoPrivate) {
-    await setRepoPrivate(page, true);
-  }
-  await expect(page.locator("#thread")).toContainText("No messages yet");
-  await page
-    .locator("#text")
-    .fill(
-      options.message ?? "<@U0BOT> please add a greet() helper and open a PR",
-    );
-  await page.locator("#send").click();
-  await expect(
-    page.locator(".msg.bot").filter({ hasText: "Add greet() helper" }),
-  ).toBeVisible();
-
-  const webLink = page.locator('.msg.bot a[href*="/agents/"]').first();
-  await expect(webLink).toBeVisible();
-  await webLink.click();
-  await expect(page).toHaveURL(/\/agents\//);
-}
-
-// The SDK hydrates an idle thread's transcript from getState on load, which can
-// briefly lag; a reload re-fetches it. Retry until the PR link renders.
-async function openMultiRepoPrThreadViaSlackLink(page: Page) {
-  await page.goto("/mock/slack");
-  await page.locator("#reset").click();
-  await page
-    .locator("#text")
-    .fill(
-      "<@U0BOT> E2E_MULTI_PR open related pull requests in both repositories",
-    );
-  await page.locator("#send").click();
-  await expect(
-    page.locator(".msg.bot").filter({ hasText: "anotherorg/companion" }),
-  ).toBeVisible();
-
-  const webLink = page.locator('.msg.bot a[href*="/agents/"]').first();
-  await expect(webLink).toBeVisible();
-  await webLink.click();
-  await expect(page).toHaveURL(/\/agents\//);
-}
-
-async function expectTranscriptVisible(page: Page) {
-  await expect(async () => {
-    await page.reload();
-    await expect(
-      page.getByRole("link", { name: "Add greet() helper" }).first(),
-    ).toBeVisible({ timeout: 8000 });
-  }).toPass({ timeout: 60000 });
-}
-
-async function waitForThreadIdle(page: Page, threadId: string) {
-  await expect
-    .poll(
-      async () => {
-        const res = await page.request.get(
-          `/dashboard/api/threads/${threadId}?mark_viewed=false`,
-        );
-        if (!res.ok()) return "unknown";
-        return ((await res.json()) as { status?: string }).status ?? "unknown";
-      },
-      { timeout: 30_000, intervals: [500] },
-    )
-    .not.toBe("running");
-}
-
-async function waitForThreadNotBusy(page: Page, threadId: string) {
-  await expect
-    .poll(
-      async () => {
-        const res = await page.request.get(`/threads/${threadId}`);
-        if (!res.ok()) return "unknown";
-        return ((await res.json()) as { status?: string }).status ?? "unknown";
-      },
-      { timeout: 30_000, intervals: [500] },
-    )
-    .not.toBe("busy");
-}
-
-async function threadState(page: Page, threadId: string): Promise<string> {
-  const res = await page.request.get(
-    `/dashboard/api/threads/${threadId}/state`,
-  );
-  if (!res.ok()) return "";
-  return JSON.stringify(await res.json());
-}
-
-async function waitForStateToContain(
-  page: Page,
-  threadId: string,
-  text: string,
-) {
-  await expect
-    .poll(() => threadState(page, threadId), {
-      timeout: 60_000,
-      intervals: [500],
-    })
-    .toContain(text);
-}
-
-async function latestPrBody(page: Page): Promise<string> {
-  const res = await page.request.get("/mock/github/data");
-  expect(res.ok()).toBeTruthy();
-  const prs = (await res.json()) as Array<{ body?: string }>;
-  expect(prs.length).toBeGreaterThan(0);
-  return prs[prs.length - 1]?.body ?? "";
-}
-
-async function openThreadActionsMenu(page: Page) {
-  const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-  expect(threadId).not.toBe("");
-  await page
-    .locator(`[data-sidebar-frame] a[href="/agents/${threadId}"]`)
-    .click({ button: "right" });
-}
-
 test.describe("Slack → web handoff (real dashboard UI)", () => {
   test("the SAME user continues the conversation in the web app", async ({
     page,
@@ -215,478 +43,12 @@ test.describe("Slack → web handoff (real dashboard UI)", () => {
     ).toBeVisible();
   });
 
-  test("keeps pull requests from multiple repositories above the composer", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openMultiRepoPrThreadViaSlackLink(page);
-
-    const companionLink = page.getByRole("link", {
-      name: "Open anotherorg/companion pull request #2",
-    });
-    await expect(async () => {
-      await page.reload();
-      await expect(companionLink).toBeVisible({ timeout: 8000 });
-    }).toPass({ timeout: 60_000 });
-
-    const strip = page.getByTestId("thread-pull-requests");
-    await expect(strip).toBeVisible();
-    await expect(
-      page.getByRole("link", {
-        name: "Open fakeorg/demo pull request #1",
-      }),
-    ).toHaveCount(0);
-    await expect(
-      page.getByRole("button", { name: "Show 1 more" }),
-    ).toBeVisible();
-
-    await companionLink.hover();
-    const hoverCard = page.getByTestId("pr-hover-card-anotherorg/companion-2");
-    await expect(hoverCard).toBeVisible();
-    await expect(hoverCard).toContainText("anotherorg/companion #2");
-    await expect(hoverCard).toContainText("Add companion integration");
-    await expect(hoverCard).toContainText("open-swe[bot]");
-    await expect(hoverCard).toContainText("main");
-    await expect(hoverCard).toContainText("add-integration");
-    await expect(hoverCard).toContainText("1 file");
-
-    await page.getByRole("button", { name: "Show 1 more" }).click();
-    await expect(
-      page.getByRole("link", {
-        name: "Open fakeorg/demo pull request #1",
-      }),
-    ).toBeVisible();
-    await expect(companionLink).toBeVisible();
-  });
-
-  test("shows live pull request health and submits actionable fixes", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openThreadViaSlackLink(page);
-    const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-    expect(threadId).not.toBe("");
-    await expectTranscriptVisible(page);
-
-    const summary = page.getByTestId("pr-summary-fakeorg/demo-1");
-    const fixButton = page.getByRole("button", { name: "Fix PR #1 issues" });
-    await expect(summary).toHaveAttribute(
-      "data-pr-tone",
-      "text-muted-foreground",
-    );
-    await expect(summary).not.toHaveAttribute(
-      "data-pr-tone",
-      "text-success-foreground",
-    );
-    await expect(summary).toContainText("draft");
-    await expect(fixButton).toHaveCount(0);
-
-    await setPullRequestHealth(page, {
-      draft: false,
-      state: "open",
-      merged: false,
-      mergeable: true,
-      mergeable_state: "clean",
-      check_runs: [],
-      statuses: [],
-      review_threads: [],
-    });
-    await page.reload();
-    await expect(summary).toHaveAttribute(
-      "data-pr-tone",
-      "text-success-foreground",
-    );
-    await expect(summary).toContainText("open");
-    await expect(fixButton).toHaveCount(0);
-    await page.waitForTimeout(1_000);
-
-    await page.route("**/pull-request-status", (route) =>
-      route.fulfill({ status: 502, json: { detail: "GitHub unavailable" } }),
-    );
-    const failedRefresh = page.waitForResponse(
-      (response) =>
-        response.url().endsWith("/pull-request-status") &&
-        response.status() === 502,
-    );
-    await page.evaluate(() =>
-      window.dispatchEvent(new Event("visibilitychange")),
-    );
-    await failedRefresh;
-    await expect(summary).toHaveAttribute(
-      "data-pr-tone",
-      "text-muted-foreground",
-      { timeout: 5_000 },
-    );
-    await summary.focus();
-    await expect(
-      page.getByTestId("pr-hover-card-fakeorg/demo-1"),
-    ).toContainText("GitHub health is unavailable");
-    await page.keyboard.press("Escape");
-    await page.unroute("**/pull-request-status");
-
-    await setPullRequestHealth(page, {
-      mergeable: false,
-      mergeable_state: "dirty",
-      check_runs: [
-        {
-          name: "unit-tests",
-          status: "completed",
-          conclusion: "failure",
-          details_url: "https://checks.example/unit-tests",
-          required: true,
-        },
-        {
-          name: "browser-e2e",
-          status: "completed",
-          conclusion: "timed_out",
-          details_url: "https://checks.example/browser-e2e",
-          required: false,
-        },
-        {
-          name: "preview-deploy",
-          status: "in_progress",
-          conclusion: null,
-        },
-      ],
-      statuses: [
-        {
-          context: "legacy/security-scan",
-          state: "error",
-          target_url: "https://checks.example/security",
-        },
-      ],
-      reviews: [
-        {
-          author: "lead-reviewer",
-          state: "CHANGES_REQUESTED",
-          body: "The fallback must preserve the original exception.",
-          url: "https://github.example/review/1",
-        },
-      ],
-      review_decision: "CHANGES_REQUESTED",
-      review_threads: [
-        {
-          path: "agent/dashboard/routes.py",
-          line: 42,
-          comments: [
-            {
-              author: "reviewer-one",
-              body: "Handle the null response before reading the payload.",
-              url: "https://github.example/discussion/1",
-            },
-            {
-              author: "author-one",
-              body: "I handled null but still need to retain the retry reason.",
-              url: "https://github.example/discussion/1-reply",
-            },
-          ],
-        },
-        {
-          author: "reviewer-two",
-          body: "Add regression coverage for the retry path.",
-          path: "tests/dashboard/test_routes.py",
-          original_line: 88,
-          url: "https://github.example/discussion/2",
-        },
-        {
-          is_resolved: true,
-          author: "reviewer-three",
-          body: "This resolved comment must not be counted.",
-          path: "README.md",
-          line: 1,
-        },
-      ],
-    });
-    await page.reload();
-
-    await expect(summary).toHaveAttribute("data-pr-tone", "text-destructive");
-    await expect(summary).toContainText("3 checks");
-    await expect(summary).toContainText("2 comments");
-    await expect(summary).toContainText("Conflict");
-    await expect(summary).toContainText("1 pending");
-    await expect(fixButton).toBeVisible();
-
-    await summary.focus();
-    const hoverCard = page.getByTestId("pr-hover-card-fakeorg/demo-1");
-    await expect(hoverCard).toBeVisible();
-    const failingChecks = hoverCard.getByTestId("pr-failing-checks");
-    await expect(failingChecks).toContainText("unit-tests");
-    await expect(failingChecks).toContainText("browser-e2e");
-    await expect(failingChecks).toContainText("legacy/security-scan");
-    const unresolvedComments = hoverCard.getByTestId("pr-unresolved-comments");
-    await expect(unresolvedComments).toContainText(
-      "Handle the null response before reading the payload.",
-    );
-    await expect(unresolvedComments).toContainText(
-      "Add regression coverage for the retry path.",
-    );
-    await expect(unresolvedComments).not.toContainText(
-      "This resolved comment must not be counted.",
-    );
-
-    await page.keyboard.press("Escape");
-    await fixButton.click();
-    const fixPrompt = "Fresh GitHub scan:";
-    await waitForStateToContain(page, threadId, fixPrompt);
-    const state = await page.request.get(`/threads/${threadId}/state`);
-    const stateText = JSON.stringify(await state.json());
-    expect(stateText).toContain("[required] unit-tests: FAILURE");
-    expect(stateText).toContain("[optional] browser-e2e: TIMED_OUT");
-    expect(stateText).toContain(
-      "The fallback must preserve the original exception.",
-    );
-    expect(stateText).toContain(
-      "Handle the null response before reading the payload.",
-    );
-    expect(stateText).toContain(
-      "I handled null but still need to retain the retry reason.",
-    );
-    expect(stateText).not.toContain(
-      "This resolved comment must not be counted.",
-    );
-    await expect(page.getByText(new RegExp(fixPrompt)).first()).toBeVisible();
-  });
-
-  // A cold load of a finished thread must hydrate from `getState()` alone. The
-  // event stream is blocked so run replay can't stand in for that read: a
-  // long-finished run has no replay left, which is what makes a broken hydrate
-  // surface as a permanently empty transcript.
-  test("a cold load renders a finished thread's transcript without run replay", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openThreadViaSlackLink(page);
-    const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-    expect(threadId).not.toBe("");
-    await waitForThreadIdle(page, threadId);
-
-    await page.route("**/stream/events", (route) => route.abort());
-    await page.goto(`/agents/${threadId}`);
-    await expect(
-      page.getByRole("link", { name: "Add greet() helper" }).first(),
-    ).toBeVisible({ timeout: 20_000 });
-    await expect(
-      page.getByText("This thread has no messages yet."),
-    ).toHaveCount(0);
-  });
-
-  test("renders Slack mrkdwn and identifies the Slack sender", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openThreadViaSlackLink(page, {
-      message:
-        "<@U0BOT> please add a greet() helper and open a PR; review *important* R&amp;D `<https://example.com/code|code docs>` <https://example.com/slack-docs|Slack docs>",
-    });
-    await expectTranscriptVisible(page);
-
-    const slackMessage = page
-      .locator('[data-message-surface="slack"]')
-      .filter({ hasText: "Slack docs" })
-      .first();
-    await expect(
-      slackMessage.getByRole("img", { name: "Slack" }),
-    ).toBeVisible();
-    await expect(slackMessage.locator("strong")).toHaveText("important");
-    await expect(slackMessage).toContainText("R&D");
-    await expect(slackMessage.locator("code")).toContainText("code docs");
-    await expect(
-      slackMessage.getByRole("link", { name: "code docs" }),
-    ).toHaveCount(0);
-    await expect(
-      slackMessage.getByRole("link", { name: "Slack docs" }),
-    ).toHaveAttribute("href", "https://example.com/slack-docs");
-  });
-
-  test("keeps sent Slack messages visible while work is folded", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openThreadViaSlackLink(page);
-    await expectTranscriptVisible(page);
-
-    const worked = page.getByRole("button", {
-      name: /^Worked(?: for .+)? · \d+ actions?$/,
-    });
-    const acknowledgement = page.getByText("On it!", { exact: true });
-    const edit = page.getByRole("button", { name: "Edited greet.py" });
-
-    await expect(worked).toBeVisible();
-    await expect(acknowledgement).toBeVisible();
-    await expect(edit).toHaveCount(0);
-
-    await worked.click();
-    await expect(edit).toBeVisible();
-    await expect(acknowledgement).toBeVisible();
-    expect(
-      await acknowledgement.evaluate(
-        (message, entry) =>
-          Boolean(
-            message.compareDocumentPosition(entry) &
-            Node.DOCUMENT_POSITION_FOLLOWING,
-          ),
-        await edit.elementHandle(),
-      ),
-    ).toBe(true);
-  });
-
-  test("expands an Edit call into a highlighted inline diff", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openThreadViaSlackLink(page);
-    await expectTranscriptVisible(page);
-
-    const worked = page.getByRole("button", {
-      name: /^Worked(?: for .+)? · \d+ actions?$/,
-    });
-    await expect(worked).toBeVisible();
-    await worked.click();
-
-    const edit = page.getByRole("button", { name: "Edited greet.py" });
-    await expect(edit).toHaveAttribute("aria-expanded", "false");
-    await edit.click();
-    await expect(edit).toHaveAttribute("aria-expanded", "true");
-
-    const inlineDiff = edit.locator("[data-diff]");
-    await expect(inlineDiff).toBeVisible();
-    await expect(
-      inlineDiff.locator('[data-line][data-line-type="change-deletion"]'),
-    ).toContainText('return "Hello!"');
-    await expect(
-      inlineDiff.locator('[data-line][data-line-type="change-addition"]'),
-    ).toContainText('return f"Hello, {name}!"');
-    await expect(inlineDiff).toHaveAttribute("data-disable-line-numbers");
-    await expect(inlineDiff).not.toContainText("normalize");
-    await expect(inlineDiff).not.toContainText("farewell");
-    await expect
-      .poll(() => inlineDiff.locator("[data-line] span").count())
-      .toBeGreaterThan(2);
-  });
-
-  test("keeps the transcript mounted after navigation and refocus", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openThreadViaSlackLink(page);
-    const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-    expect(threadId).not.toBe("");
-    await waitForThreadIdle(page, threadId);
-
-    await page.getByRole("link", { name: "New Thread" }).click();
-    await expect(page).toHaveURL(/\/agents\/?$/);
-    await page.goBack();
-    await expect(page).toHaveURL(new RegExp(`/agents/${threadId}$`));
-    await expect(
-      page.getByRole("link", { name: "Add greet() helper" }).first(),
-    ).toBeVisible();
-
-    const foregroundHydration = page
-      .waitForRequest(
-        (request) => {
-          const path = new URL(request.url()).pathname;
-          return (
-            request.method() === "GET" &&
-            path === `/dashboard/api/threads/${threadId}/state`
-          );
-        },
-        { timeout: 1_000 },
-      )
-      .then(
-        () => true,
-        () => false,
-      );
-    await page.evaluate(() =>
-      document.dispatchEvent(new Event("visibilitychange")),
-    );
-    expect(await foregroundHydration).toBe(false);
-    await expect(
-      page.getByRole("link", { name: "Add greet() helper" }).first(),
-    ).toBeVisible();
-
-    await typeIntoComposer(page, "Can you also add a docstring?");
-    await expect(
-      page.getByText(/anything else you'd like changed/),
-    ).toBeVisible();
-    await expect(
-      page.getByRole("link", { name: "Add greet() helper" }).first(),
-    ).toBeVisible();
-  });
-
-  test("does not expose the originating Slack thread for public repos", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openThreadViaSlackLink(page);
-
-    await openThreadActionsMenu(page);
-    await expect(page.getByText("Open Slack thread")).toHaveCount(0);
-    await expect.poll(() => latestPrBody(page)).not.toContain("Slack thread");
-  });
-
-  test("exposes the originating Slack thread for private repos", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openThreadViaSlackLink(page, { repoPrivate: true });
-
-    await openThreadActionsMenu(page);
-    const sourceItem = page.getByText("Open Slack thread");
-    await expect(sourceItem).toBeVisible();
-    const popupPromise = page.waitForEvent("popup");
-    await sourceItem.click();
-    const popup = await popupPromise;
-    await expect(popup).toHaveURL(/\/mock\/slack/);
-    await expect.poll(() => latestPrBody(page)).toContain("Slack thread");
-  });
-
-  // The queued card is optimistic, so a regression shows up as a flash the DOM
-  // holds for only the length of one request — too short for a locator poll.
-  test("never flashes a queued card when no run is in progress", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openThreadViaSlackLink(page);
-    const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-    expect(threadId).not.toBe("");
-    await waitForThreadIdle(page, threadId);
-    // The dashboard's status can report a finished run before LangGraph drops
-    // the thread out of `busy`, and `busy` is the exact condition the queue
-    // endpoint accepts on. Wait for it, or the send legitimately queues.
-    await waitForThreadNotBusy(page, threadId);
-
-    await page.evaluate(() => {
-      const seen = { value: false };
-      (window as unknown as Record<string, unknown>).__queuedCardSeen = seen;
-      new MutationObserver(() => {
-        if (document.querySelector('[data-testid="queued-message"]'))
-          seen.value = true;
-      }).observe(document.body, { childList: true, subtree: true });
-    });
-
-    await typeIntoComposer(page, "Can you also add a docstring?");
-    await expect(
-      page.getByText(/anything else you'd like changed/),
-    ).toBeVisible();
-
-    const flashed = await page.evaluate(
-      () =>
-        (
-          (window as unknown as Record<string, unknown>).__queuedCardSeen as {
-            value: boolean;
-          }
-        ).value,
-    );
-    expect(flashed).toBe(false);
-  });
-
   test("keeps follow-ups visible across the queued-to-transcript handoff", async ({
     page,
   }, testInfo) => {
     await loginAs(page, SAME_USER);
     await openRunningThreadViaSlackLink(page);
-    const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-    expect(threadId).not.toBe("");
+    const threadId = threadIdFromUrl(page);
 
     const queuedText = "Please queue this follow-up while you finish the PR.";
     const busyComposer = composerFor(page, /Send a message to queue next/);
@@ -737,8 +99,7 @@ test.describe("Slack → web handoff (real dashboard UI)", () => {
     const prompt = "list my open langchainplus PRs";
     await typeIntoComposer(page, prompt);
     await expect(page).toHaveURL(/\/agents\/[^/]+$/);
-    const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-    expect(threadId).not.toBe("");
+    const threadId = threadIdFromUrl(page);
 
     const userMessage = page
       .getByTestId("user-message")
@@ -774,8 +135,7 @@ test.describe("Slack → web handoff (real dashboard UI)", () => {
     await editor.pressSequentially("first sender payload message");
     await editor.press("Enter");
     await expect(page).toHaveURL(/\/agents\/[^/]+$/);
-    const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-    expect(threadId).not.toBe("");
+    const threadId = threadIdFromUrl(page);
     await waitForStateToContain(
       page,
       threadId,
@@ -921,7 +281,9 @@ test.describe("Slack → web handoff (real dashboard UI)", () => {
         ),
       )
       .toBe(true);
-    await page.waitForTimeout(5_000);
+    // The flash this guards against lands within a navigation, so a couple of
+    // seconds of sampling is enough to catch it.
+    await page.waitForTimeout(2_000);
 
     const observations = await page.evaluate(
       () =>
@@ -938,112 +300,6 @@ test.describe("Slack → web handoff (real dashboard UI)", () => {
     );
     expect.soft(observations.messageDisappeared).toBe(false);
     expect.soft(observations.newChatReturned).toBe(false);
-  });
-
-  test("renders a web follow-up exactly once", async ({ page }) => {
-    await loginAs(page, SAME_USER);
-    await openThreadViaSlackLink(page);
-    const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-    expect(threadId).not.toBe("");
-    await waitForThreadIdle(page, threadId);
-    await waitForThreadNotBusy(page, threadId);
-
-    const followUp = "Can you also add a docstring?";
-    await typeIntoComposer(page, followUp);
-    // The agent's canned reply is already in the transcript from the Slack run,
-    // so wait on the run persisting this message rather than on any reply text.
-    await waitForStateToContain(page, threadId, followUp);
-    await waitForThreadIdle(page, threadId);
-    await waitForThreadNotBusy(page, threadId);
-
-    await expect(
-      page.getByTestId("user-message").filter({ hasText: followUp }),
-    ).toHaveCount(1);
-  });
-
-  test("renders structured input envelopes safely and keeps legacy messages", async ({
-    page,
-  }) => {
-    await loginAs(page, SAME_USER);
-    await openRunningThreadViaSlackLink(page);
-    const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-    expect(threadId).not.toBe("");
-    await waitForThreadIdle(page, threadId);
-
-    await page.route(
-      `**/dashboard/api/threads/${threadId}/state`,
-      async (route) => {
-        const response = await route.fetch();
-        const body = (await response.json()) as {
-          values?: { messages?: Array<Record<string, unknown>> };
-        };
-        const messages = body.values?.messages ?? [];
-        body.values = {
-          ...body.values,
-          messages: [
-            {
-              type: "human",
-              id: "entity-person",
-              content:
-                '<dynamic-context kind="person" id="github:alice"><display_name>Alice</display_name></dynamic-context>',
-            },
-            {
-              type: "human",
-              id: "entity-system",
-              content:
-                '<dynamic-context kind="system" id="system:scheduler"><display_name>Scheduler</display_name></dynamic-context>',
-            },
-            {
-              type: "human",
-              id: "structured-person",
-              content:
-                '<input-message sender="github:alice" surface="web" kind="human"><content>Person says &lt;img data-e2e-injected src=x&gt;</content></input-message>',
-            },
-            {
-              type: "human",
-              id: "structured-system",
-              content:
-                '<input-message sender="system:scheduler" surface="automation"><content>Automation checks CI</content></input-message>',
-            },
-            {
-              type: "human",
-              id: "legacy-e2e",
-              content: "Legacy stays visible",
-            },
-            ...messages,
-          ],
-        };
-        await route.fulfill({ response, json: body });
-      },
-    );
-
-    await page.reload();
-    await expect(
-      page.getByText("Person says <img data-e2e-injected src=x>"),
-    ).toBeVisible();
-    await expect(page.locator("img[data-e2e-injected]")).toHaveCount(0);
-    await expect(page.getByText("Automation checks CI")).toHaveCount(0);
-    const systemChip = page.getByRole("button", { name: "Scheduler" });
-    await expect(systemChip).toBeVisible();
-    await systemChip.click();
-    await expect(page.getByText("Automation checks CI")).toBeVisible();
-    await expect(page.getByText("Legacy stays visible")).toBeVisible();
-    await expect(page.getByText("github:alice", { exact: false })).toHaveCount(
-      0,
-    );
-    await expect(
-      page.getByText("system:scheduler", { exact: false }),
-    ).toHaveCount(0);
-    await expect(
-      page
-        .locator('[data-message-sender-kind="person"]')
-        .filter({ hasText: "Person says" }),
-    ).toBeVisible();
-    await expect(
-      page
-        .locator('[data-message-sender-kind="system"]')
-        .filter({ hasText: "Automation checks CI" }),
-    ).toBeVisible();
   });
 
   test("stops a Slack-started run from the web app", async ({ page }) => {
@@ -1134,8 +390,7 @@ test.describe("Slack → web handoff (real dashboard UI)", () => {
   }) => {
     await loginAs(page, OTHER_USER);
     await openThreadViaSlackLink(page);
-    const threadId = new URL(page.url()).pathname.split("/").pop() ?? "";
-    expect(threadId).not.toBe("");
+    const threadId = threadIdFromUrl(page);
 
     // The same thread + transcript is visible…
     await expectTranscriptVisible(page);
