+++ tests/e2e/tests/full_flow.spec.ts
@@ -74,12 +74,6 @@ test.describe("Open SWE full flow", () => {
       .filter({ hasText: "Add greet() helper" });
     await expect(reply).toBeVisible();
     await expect(reply.locator('a[href*="/pull/"]')).toBeVisible();
-
-    await page.goto("/mock/github");
-    await expect(page.locator('.pr[data-pr="1"]')).toContainText(
-      "Add greet() helper",
-    );
-    await expect(page.locator('.pr[data-pr="1"]')).toContainText("greet.py");
   });
 
   test("Slack breakout request starts a new top-level Open SWE thread", async ({
@@ -112,25 +106,5 @@ test.describe("Open SWE full flow", () => {
         .locator(".msg.bot")
         .filter({ hasText: "I started a separate Open SWE thread" }),
     ).toBeVisible({ timeout: 60_000 });
-  });
-
-  test("a message that does not mention the bot produces no run and no PR", async ({
-    page,
-  }) => {
-    await page.locator("#mention").uncheck();
-    await page
-      .locator("#text")
-      .fill("just chatting with the team, nothing for the bot");
-    await page.locator("#send").click();
-
-    await expect(
-      page.locator(".msg").filter({ hasText: "just chatting" }),
-    ).toBeVisible();
-    // No agent activity: give the (non-)run a moment, then assert nothing came back.
-    await page.waitForTimeout(3000);
-    await expect(page.locator(".msg.bot")).toHaveCount(0);
-
-    const prs = await (await page.request.get("/mock/github/data")).json();
-    expect(prs.length).toBe(0);
   });
 });
